Additions (#2)
* Weighted Median

This code is written as a function, it can be added to existing code. This code calculates the weighted median of a list of scores and their corresponding weights. The code will then prompt users if they would like to see the calculation results and also the graph of the data.

* Update README.md

* Update README.md

* Update README.md

* Update README.md